requirements.txt
# --- Core ---
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# --- LLM Access ---
dashscope>=1.14.0     # Qwen (Dashscope)
openai>=1.6.0         # OpenAI 官方 API 和兼容 API
requests>=2.31.0      # Ollama API 调用

# --- Utilities ---
rich>=13.7.0
loguru>=0.7.2

# --- Java Interop ---
jpype1>=1.4.0

# --- CLI ---
typer>=0.9.0

# --- 可选：密钥环存储 API Key（否则仅用环境变量 / .env）---
# keyring>=24.0.0

# --- Dev / Debug ---
tqdm>=4.66.0
pytest>=7.4.0